error check function 1

+++ push_swap/srcs/ft_atoi.c
@@ -6,16 +6,16 @@
 /*   By: jung-lee <jung-lee@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/17 15:00:37 by jung-lee          #+#    #+#             */
-/*   Updated: 2021/06/01 17:50:12 by jung-lee         ###   ########.fr       */
+/*   Updated: 2021/06/29 13:39:41 by jung-lee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/push_swap.h"
 
-int	ft_atoi(const char *str)
+long long		ft_atoi(const char *str)
 {
-	int	n;
-	int	sign;
+	long long	n;
+	int			sign;
 
 	while (*str == '\n' || *str == '\t' || *str == ' '
 			|| *str == '\v' || *str == '\f' || *str == '\r')
